put stuff where it belongs
